fix error handling on Github authorization

+++ web/app/controllers/endpoints_controller.rb
@@ -345,16 +345,18 @@ class EndpointsController < ApplicationController
         'endpoint_name' => @endpoint.name,
     }
 
-    session[:prev_uri] = request.referer
+    session[:prev_url] = request.referer
 
     redirect_to '/auth/github'
   end
 
   def after_authorization
     issue_info = session[:issue_info]
+    return redirect_to root_path unless issue_info
+
     @issue = GithubIssue.new(description: issue_info['description'], title: issue_info['title'])
     @issue.save(issue_info['endpoint_name'], session[:oauth_token])
-    prev_uri = issue_info['prev_uri']
+    prev_url = session[:prev_url] || root_url
 
     GithubHelper.revoke_oauth_token(session[:oauth_token])
     session.delete(:issue_info)
@@ -362,7 +364,7 @@ class EndpointsController < ApplicationController
 
     if @issue.errors.any?
       @success = false
-      redirect_to prev_uri, flash: {failure: "Failure on Issue Creation: \n #{@issue.errors.full_messages.join("\n")}"}
+      redirect_to prev_url, flash: {failure: "Failure on Issue Creation: \n #{@issue.errors.full_messages.join("\n")}"}
     else
       @success = true
       redirect_to "https://github.com/#{Rails.application.secrets.github_repo}/issues/#{@issue.id}"
